Extract convert_to_crash_data
This utility will be useful for Socorro code to convert between its data
structures to the siggen one.
